requirements.txt
pandas
numpy
scikit-learn
beautifulsoup4
selenium
requests
matplotlib
seaborn
jupyter
amadeus
python-dotenv
